Added a CMake module to find the Event Transfer library, and to then compile the online options into the analyzer. Note that there is exactly one copy of the ET library which is correctly compiled on the Hall A cluster right now, and the "easy to find" copies which are in the CODA directores are not 64-bit compiled like everything else, so the "FindET" is restricted to a single path.
Minor changes in the ET support classes were needed to use this 64-bit
ET library to connect to the CODA distribution's 32-bit compiled ET
server.
In particular, we open the ET station as "remote" using
   et_open_config_setmode(openconfig, ET_HOST_AS_REMOTE);

A config file to enable online mode is included; you should use both
this and the config file defining your analysis options as:
   ./build/qwparity  --config prexinj.conf  --add-config online_apar.conf

The online mode seems to require "tree trimming" to be active; I have
added an initial tree_trim.in file, but it will certainly need to be adapted.