Dockerfile and travis-ci config to build japan in the cloud.
Dockerfile builds a docker image based on jlabce:2.2. Adds boost
libraries and compiles. Running syntax:
  docker run jeffersonlab/japan 'qwparity -h'

.travis.yml just instructs travis-ci.org to build the docker image
and push it to docker hub at
  https://hub.docker.com/r/jeffersonlab/japan/

Since a docker image is already on docker hub, the following commands
should already work:

If you have root access or are in the right group (typically docker)
  docker run jeffersonlab/japan 'qwparity -h'
or in singularity for userspace (e.g. on ifarm or institutional systems)
  singularity run docker://jeffersonlab/japan:latest