Made QwIntegrationPMT and QwCombinedPMT templated in fADC class
This means we now need to specify QwIntegrationPMT<QwVQWK_Channel>
instead of merely QwIntegrationPMT. The reason for this change is
to be able to use other ADC types as integrated PMT channels, which
in turn is needed for HAPPEX ADC18 channels (even if only for testing
with the mock data generator).

All detector classes that use QwIntegrationPMT or QwCombinedPMT have
been updated to using QwIntegrationPMT<QwVQWK_Channel> and equiv.
It is up to the detector classes to support the different ADC channel
types in their channel map parsing routines, if desired.

Additional work needed:
- resolving the question of whether we ever will combine different
  ADC types in a QwCombinedPMT,
- adding a virtual VQwIntegrationPMT that is not templated but can
  be used as return type for some of the functionality,
- remove the -fpermissive compiler flag, which will require that we
  look at the SetSingleEventCuts function and whether it should have
  default arguments in the overloaded version in the inherited classes.