Changes to allow bank decoding as implemented in Hall A DAQ.
By default, if the subbank IDs are 31 or less, they would be interpreted as
ROC IDs instead.  A new command-line/configuration-file flag has been added
to allow low bank ID values.  If that flag is set, all ROC banks MUST be
built from sub-banks, but then the sub-banks may have any ID value.
Subbank IDs cannot be repeated within the same ROC, but they can be repeated
in different ROCs.

To allow this to work, the configuration file will be re-parsed at the
start of each run, so if the new flag is different for different run
ranges, we can specify it in an "extra" config file that can be run-ranged.