Decode and report SIS3801 header bits.
The header bits for the SIS3801 scalers will now be decoded and reported in
the raw tree.  The header is not bit shifted, and does not have the channel
number extracted.  If you just want to see the SIS3801 user bits in the
header, you would do something like:  ((header & 0xc0000000)>>30).