Cmake changes to almost get the feedback analyzer working.
Added a modeule to find the EPICS libraries on the adaq cluster,
and a CMakeList to compile the Feedback library if both EPICS and ET have
been found.
I wasn't able to get the QwFeedback library and the feedback executable
to both be compiled and linked though.
Maybe @wdconinc can look at it...