Temporary root file names will now be "probably unique"
I have reverted the rootfile class to use a probably unique
temporary root file name instead of using "test001.root".
This lets us be generating multiple rootfile simultaneously.

The file names will be like:
	prexALL_1034.adaq3.jlab.org.25115.root
with the machine name and process ID as the 3rd and 4th pieces
of the name.
On successful completion, the file will be renamed like:
	prexALL_1034.root