Cleanup of naming of data handler derived classes.
The data handler derived classes had been based on older code which we had referred to as "regression" classes, and they had internal and external variables with variations of "regression" in the name.  Since we actually call these classes "correlator", "combiner", or "corrector", I've tried to go through the code and make the associated variables closer in name to the classes as used.