Fix to "online" mode so it doesn't wait forever to see the first EPICS event.
If we are in "online" mode, the engine will skip the initial attempt to find
an EPICS event, and just begin the main event processing loop right away.

This was a problem for us, because the PREX data stream has no EPICS events yet, so we'd never actually decode any events from the ET system.