Load both New and Gui ROOT libraries, first and in that order.
Added required FindROOT module Gui without which the New library
does not work for JLab CE 2.3. Changed the order of linker libraries
such that the specified additional ROOT_FIND_COMPONENTS are added
at the beginning of the list, not the end, as required for New.

Maybe of interest to @hansenjo too, since a modifcation of a cmake
module from Podd.