Changes in QwOptions to allow "add-config" to be used on both the command line and in a configuration file.
The option "add-config" is now defined with the "composing" flag, meaning that all instances will be used by the options parser.
When adding a new config file, test to be sure we haven't added it to the list already.
After parsing a config file, check to see if a new config file has been added.

NOTE:  This will allow configuration files to be added from either within a configuration file, or from the command line, but
if the options within those files are in conflict, the earliest one will be taken, and if the commandline options conflict with
any of the configuraiton flags, the commandline will be taken.
Use the "add-config" only for options that are appearing uniquely in their config files.

DO NOT use "set-config" from within a config file.  The results would be not what you meant.