Deleting unneccessary macros that were just a stab at doing too much in a too complicated way. Cameron has copied them to his own work-space and will pursue the macro-as-generic plotter technique developed in exampleMacro2.C instead.
APEX has some nice macros that show how to grab a run-number from
environment variables in case the macro is being run from within
panguin and the run-number integer needs to be obtained for log
file printing or for more convenient legend production. This feature
can be added at a later time, and requires using shell wrapper scripts
that take the -r parameter and set an environment variable with it.