Modified the stability cut functionality.
Previously (see discussion in issue #92) the stability cut was only
evaluated when we reached the end of the event ring.  Now the stability
cut will be evaluated as each event is pushed into the ring.
This improves the detection of beam trips.

However, there's still an issue with recognizing the start of the
beam trip and really cutting a full "ring.size" events before the trip.
This version does definitely cut away before the beam trips, but it
is fewer events than it ought to be.

The next stage should refactor how we do the stability check so that
we look at the width and at the values of the event which is being
pushed into the ring.  A failed single-event cut on a channel we are
monitoring stability should also trigger the stability cut.
Possibly a single-event change which is "large" should also trigger
the stability cut.

Due to the functionality change in the event cut, I also changed the
stability cut settings in the CHA beamline files.