Modified blinder to not print counters on destruction if all counters are zero.
The blinder object willususally print the counters for the number of patterns (or pairs) which have passed
or failed the blinder checks.  However, if a copy of the blinder has not been used to do any blinder checks
(such as from a temporary un-used copy) it will not print anything during deconstruction.