Initalize all data handlers from a parameter file instead of command line options.
The data handlers in the data handler array are now initialized using a parameter file.
There should now be no hardcoded calls to specific datahandler objects in the main routine,
or in the data handler array.

If a specific type of data handler requires a new parameter file entry, see the "LoadDetectorMaps" in the derived classes for how to extend the parameter file decoding.

This has been tested to compile and execute on run 1460, but the outputs have not been checked for consistency with previous versions.