Removed default "datahandlers" value, so that you must specify a datahandler in the config file.
The default will be to not have an active data handler array, so if you want to use the data handlers, you need to specify the handler arry parameter file in your configuratin file.

Added "prex_datahandlers.map" to the prex.conf file.  With this in the configuration file, you can bypass it by specifying '--datahandlers ""' on the command line.