Shifted the copy constructor from QwCombiner to VQwDataHandler.
The copy constructor as defined in QwCombiner was actually appropriate
fot eh virtual base class of the data handlers.  By shifting it into
the base class, all data handlers will have a copy constructor defined,
allowing the copy for accumulating the runnign sum to be made.

In this version, both QwCombiner and LRBCorrector objects can have
a running sum calculated.