Updating the cavity BPM decoder to work with the PREX data.
The cavity BPMs used so far in PREX have three input channels which are: X*current, Y*current, and current.  This set of changes allows all three to be decoded and the actual positions to be computed by taking the ratio of the position dependent channels and the current channel.

Right now, the input channels are labeled as "X", "Y", and "Q", and the output positions are labeled as "AbsX" and "AbsY".  We may want to change the inputs to be "XI" and "YI" (or XQ & YQ), and hte output positions as "X" and "Y".  Or to change the processing so that the conversion to position is done directly in the variable which holds the input data.

Right now the "asym" values for the position dependent input channels are computed as differences.  We should decide if that is what we want, or if we should not compute pair/pattern dependent quantities of the position dependent input channels at all.  Note that in the stripline BPM processing, we do not create pair/pattern yields & asymmetries of the individual wires, so we might drop the XI and YI channels out of "symmetry".