Added parameter files with the RHRS channels commented out.
Beginning with run 1968, the RHRS VQWK trigger was disconnected due to the cables being tangled with others being removed.
The crate was still participating in the DAQ, but the data is all empty for the RHRS main detector VQWK and the SAMs.

The SAMs and the RHRS main detector channels have been commented out in their channel maps, so we don't get spurious error
counts from those modules.

@cameronc137 Once the RHRS trigger has been reestabilshed, close the run range for these two files.