Added explicit "not empty" checks to all QwDataHandlerArray member functions.
Perhaps some of these were not strictly needed, for functions which are just a loop
over the array elements (and so ought to do nothing if there are no array elements),
but some functions threw seg faults, if they tried to access other parts of the
class structure.
The specific problem was trying to use the data-source pointer if we hadn't
initalized it.