Introduce a new mixin class MQwStoreObjects
The idea is that there is no real need to reimplement or redefine
a lot of ConstructObjects functions at every level. We can do this
with a mixin class (like we already do, partially, for histograms).

This removes ConstructObjects again (after only 5 days...) and puts
the whole shebang in one single mixin. So:

`QwHelicityPattern: public MQwStoreObjects` contains a fBlinder
which is `QwBlinder: public MQwStoreObjects`. In the constructor
of QwHelicityPattern we explicitly set QwBlinder as a dependent
for purposes of constructing and storing objects. The loops are
then done in/by the mixin. Then it's mixins all the way down.

How about QwSubsystemArray, you ask? We can likely use STL traits
and clever template magic to automatically loop over the contents.

How about trees, histograms? Can something similar be done there?
Maybe. It sure would make a lot of our root interface code simpler...