Minor changes to the blinder.
Changed the example blinder map file to begin at run 0 and be open ended.

Modifed the blinder option names and descriptions to give clearer indication how to force the blinder to be inactive or to be active when the EPICS system is not able to determine the blindability of the target.