added event datahandler, extra math in bevington, and made the corrector read a variable map file.
had to change qwdatahandlerarray to qwdatahandlerarrayHEL and EVT to start crating the event level datahandler. right now they are essentially copies of eachother.
The LinReg_Bevington_Pebay.cc now calculates the corrected Y values, and the variances on the corrected valus and prints them to the screen.
LRBCorrector.cc now reads in variable names from a map file in the same way QwCorrelator.cc does.