Helper function VQwHardwareChannel::SetDataToSaveByPrefix
We decide what to store for each VQWK data element based on the name of
the prefix used when we construct the histograms (don't ask). If we want
to customize what we store (e.g. second moments), that's the existing
functionality we have. So, we use this to store second moments when the
prefix contains "stat_". We continue to store only derived (not raw)
values when the prefix contains "asym_", "yield_", or "diff_".

Ideal? No. But I tried 7 other 'elegant' ways first and none worked.