Added QwEventBuffer option to set the ET system wait mode.
The analyzer's ET cliuent wait mode can be set to 0 or 1.
When it is zero, the analyzer will "wait forever" for new events.
When it is one, the ET client will time out (time out period is
defined in the evio library).

This was the start of getting the online analyzer to clear
the ROOT output on run transitions.