Additions to LRB functionality: store IV/DV correlations
Now this also determines and stores all correlation matrices instead of
only the IV correlations. Similarly it now also stores all covariance
matrices.

Small impact: in the blueR####new.slope.root files the matrix
IV_rawVariance is now called IV_IV_rawVariance and contains the
covariance matrix of IV against IV. The new variable IV_rawVariance is
the column of variances of each IV with itself. This is in analogy with
DV_rawVariance which is the column of variances of each DV with itself.