Fix accumulation of VQWK channel
Previously, accumulation of n2 > 1 good events into a channel (which
already had n1 good events) would fail (though not originally and not in
the other hardware channels). Now it does not fail anymore.

If n2 = 0 and the error flag is the event we're adding is good, then we
assume that n2 really should have been 1 (essentially this is the good
event check) and we accumulate the event.

If n2 = -1 and the error flag EXCEPT stability is good, then we
deaccumulate the event, otherwise n2 = 0. (Previously those were the
only allowed options.)

If n2 = 0 we don't do anything.
If n2 = -1 and we don't empty the running sum, then we deaccumulate a
single event.
If n2 = -1 and we empty the running sum, then we complain.
If n2 = -1 and we go to negative number of events, then we complain
loudly.
If n2 = 1 then we accumulate a single event (simpler math).
If n2 > 1 then we accumulate a bunch of events (more complicated math).