Unwind event ring after doing a full loop
Nothing is actually done with the events, but optionally (commented)
there is a line to print out the residual event ring rolling average
after unwinding. This should be zero and if not it indicates an error in
how the event ring or accumulation is handled.

Conclusions: not zero for scalers (they don't properly deaccumulate ->
don't put event ring stability cuts on scaler channels).