Changes to how pairs/patterns with blinding errors appear in ROOT files.
For pairs/patterns that have a blinding failure, blindable asymmetries
will be set to a value of -1 in natural units (1.e. -1000000 ppm)
instead of zero.

Pairs/patterns with a blinding failure will have the ErrorFlag set to
include the blinding error bit instead of just having that appear in
the individual device error codes.
The blindable asymmetries already had their individual error codes
marked in the case of blinding errors, but this now will have the
entire pair/pattern so marked.