Blinder changes and parameter files to support run-ranged switching between PREX and CREX blinding settings.
The blinding maps now can have a boolean flag indicating if we should be using the CREX target position list or not; it defaults to false.  The only blindable position for CREX-mode is the Ca-48 location, with encoder position range between 14.5e6 and 18.0e6, as extracted from HALOG 3736585.  This puts the limits halfway between that target and its neighbors.

Closed the PREX blinder map file at run 5103, which was the last injector parity scan on September 15.

Created an interim blinder mapfile from runs 5104-5274, starting on October 3 and going to November 7.  This uses the CREX blinding box but is a different seed.  This ought to not result in any of these runs being blinded, but just in case.

Created the CREX blinder map, starting at run 5275, which as of this moment has not been taken.