Modifications to allow LRB corrections using the burst-level slopes.
The burst index is passed to the burst-level datahandler array when we start a new burst.  The LRBCorrector can then use it to look up the correct "cycle" from the slopes file.

The burst counter is included in the mul tree and the burst tree.

I checked that if a run has only one burst, the run-level and burst-level slopes are identical, as are the corrections.