Removing unneeded "Enable/Disable" functions from QwHelicityPattern.
This largely resolves issue #159.

The QwHelicityPattern class is in charge of setting and using the flags to enable/disable analysis for:
    AlternateAsymmetry
    Burst
    RunningSum
    Difference
    Pairs
The QwHelicityPattern class had "Enable/Disable" functions and "Is*Enabled" functions for all of these, but almost none of them were meaningfully used.  I have removed almost all of them.
The "DisablePairs" function remains, as it is used when creating the summing copies of the QwHelicityPattern class for the full run and for bursts.
I have also temporarily left a commented copy of the "IsBurstSumEnabled" function.