Added commentary in the edittree method to guide future slope blessing procedure.
We also need a bit of code to determine from the data what the BMOD trim offset values should be (instead of hard coding into the input.txt file)

Also we need a script for the respin that will decide which input.txt file (and set of coils and detectors) should be used.

We should make sure the code works nicely for getting plots out quickly (maybe even in panguin) and for doing different permutations of BPM/coil sensitivity combinations quickly.