Changed which Hll A Aq is reported to EPICS and added EPICS reporting of the Hall C Aq.
The EPICS variables BCM1H01_AsymMean and BCM1H01_AsymWidth will now represent whatever we are using as "q_targ", which is also the asymmetry we are doing the feedback upon.  This does mean that if we're doing injector charge feedback, this EPICS variable is not "BCM1H01".

The EPICS variables BCM3H04_AsymMean and BCM3H04_AsymWidth will be filled with the asymmetry of channel hallc_bcm_4c.