Hotfix to allow selecting run-dependent blindable target positions.
This extends the mechanic that we used to differentiate between the PREX and
CREX taret position ranges to allow more than two sets of blindable target
ranges.
The PREX positions are associated to target index "-1", and the CREX positions
to date are index "1".  The new CREX target encoder position will be "2", and we
can easily define more later if needed (like if the encoder changes).