Fix for issue #269:  incorrect labels for position differences in summary trees
The routines for labeling the position difference variables in ROOT trees have
been corrected so that the "diff" label is used instead of the "asym" label in
summary trees as well as the multiplet trees.

This has no change in the calculations, it is only impacting the tree branch
names.