Changed the global BCM cuts to use "bcm_target" for PREX2.
The global BCM cuts (which are used to form the beam trip cuts) have
been changed to use the combined BCM variable "bcm_target" for the
entire PREX2 running period.  This variable is usually defined to be
(2*bcm_an_us + bcm_an_ds)/3, except for the periods where the analog
BCMs were not working, when it is instead defined to be bcm_dg_ds.